+++ plugin_ide.ui/src/com/googlecode/goclipse/ui/preferences/GoToolsPreferencePage.java
@@ -10,28 +10,23 @@
  *******************************************************************************/
 package com.googlecode.goclipse.ui.preferences;
 
-import static org.eclipse.jface.layout.GridDataFactory.fillDefaults;
 import melnorme.lang.ide.ui.LangUIPlugin_Actual;
 import melnorme.lang.ide.ui.tools.AbstractDeamonToolPrefPage;
 import melnorme.util.swt.SWTFactoryUtil;
 
-import org.eclipse.core.runtime.IPath;
 import org.eclipse.jface.layout.GridDataFactory;
 import org.eclipse.jface.layout.GridLayoutFactory;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Group;
-import org.eclipse.swt.widgets.Label;
 import org.eclipse.ui.IWorkbenchPreferencePage;
 
 import com.googlecode.goclipse.core.GoToolPreferences;
-import com.googlecode.goclipse.core.tools.GocodeServerManager;
 
 public class GoToolsPreferencePage extends AbstractDeamonToolPrefPage implements
 		IWorkbenchPreferencePage {
 	
 	protected Group oracleGroup;
-	protected Label gocodeLabel;
 	
 	@Override
 	protected void doCreateContents(Composite block) {
@@ -53,14 +48,6 @@ public class GoToolsPreferencePage extends AbstractDeamonToolPrefPage implements
 	@Override
 	protected void createDaemonPathFieldEditor(Group group) {
 		super.createDaemonPathFieldEditor(group);
-		IPath gocodePath = GocodeServerManager.getDefaultGocodePath();
-		String labelText = 
-				gocodePath == null? "" : 
-				"If path is empty, the built-in gocode will be used:\n" + gocodePath.toOSString();
-		
-		gocodeLabel = SWTFactoryUtil.createLabel(group, SWT.WRAP, 
-			labelText,
-			fillDefaults().span(3, 1).grab(true, false).indent(10, 0).minSize(200, SWT.DEFAULT).create());
 	}
 	
 }