+++ documentation/ChangeLog.md
@@ -10,6 +10,7 @@
  * Fixed #112: InvalidPathException invoking content assist. 
   * This also fixes potential performance problems invoking content assist.
   * However the documentation popup information will no longer be show for any content assist entries.
+ * Removed bundled `gocode` as it was very out of date (closes #88).
 
 ### Goclipse 0.9.1
  * Fixed: Builder does not ignore folders with names starting with "_" or ".".